Provide more context when starting PostgreSQL
Currently, if a connection to the PostgreSQL database cannot be
established, operators are instructed to wait indefinitely. They need to
exercise their own judgement to determine that an unrecoverable error
may have occurred, and they must then formulate their own plan to
diagnose the problem. This change introduces a suggestion about the
possibility of failure and some contextual information that will assist
in debugging. It also causes the script to fail immediately for truly
exceptional circumstances such as the absence of required modules.