Prevents endless failing Stripe requests
When a database record has a stored customer_id that can't be found in Stripe (mainly due to copying data from one environment where we're using live Stripe or a sandbox to another where we're not), we would endlessly try to fetch the missing record. This clears the customer_id after a retreival error so a new customer can be created in whatever Stripe environment is being used.