Eliminates Wix Sync race conditions
In a few cases, a task to sync an org with Wix could execute before the organization and its members are done saving to the database. This would leave the task unable to find the organization it needs to sync. By waiting until a commit, the we ensure the task has the data it needs to succeed.

This also addresses a separate, low-risk  race condition where the active subscription count could change between retreiving it and starting a subscription, leading us to exceed our maximum active subscription count.